Refactor to rename and streamline feature management components
Renamed `unified_feature_manager.py` to `feature_manager.py` for clarity and updated references accordingly. Improved code formatting and organization, including refactoring multiline statements for readability. Cleaned up unnecessary whitespace and redundant code, ensuring consistency across map component handling.
