+++ src/logs/neoworldbuilder.log
@@ -729,3 +729,4 @@
 {"event": "Configuration loaded successfully", "timestamp": "2025-05-24T05:22:59.225273Z", "level": "info"}
 {"event": "Configuration loaded successfully", "timestamp": "2025-05-24T05:34:02.533259Z", "level": "info"}
 {"event": "Configuration loaded successfully", "timestamp": "2025-05-24T05:35:06.188784Z", "level": "info"}
+{"event": "Configuration loaded successfully", "timestamp": "2025-05-24T05:37:20.766443Z", "level": "info"}
