+++ src/ui/components/map_component/feature_manager.py
